[Core] Fixes: The form in a summary element disappears.
Problem:
--------

If you save a summary form and try to edit it a second time, the form slides down, and
immediately slides up again. Even the summary view is not shown anymore. This only happens
under certain circumstances (in Applications, it is working; in Jobs it isn't)

Cause:
------

The click event listener on the "edit"-Button is attached twice, when no content is returned
in the ajax result, because the old content of the summary view is not replaced in that case.

Solution:
---------

Only attach the listener, if content is returned in the ajax call.

(refs #5623)